Made a new file for SecondView
